move vaccines components to their own folder
