+++ src/components/Rules/RulesContainer/index.js
@@ -5,7 +5,7 @@ import AddRuleButton from '../AddRuleButton';
 import SaveChangesButton from '../SaveChangesButton'
 import TestRulesButton from '../TestRulesButton'
 import SimulateRulesButton from '../SimulateRulesButton'
-import EditVaccinesButton from '../EditVaccinesButton'
+import EditVaccinesButton from '../../Vaccines/EditVaccinesButton'
 import { Grid } from '@material-ui/core';
 import useStyles from './styles';
 import NoRulesMessage from '../NoRulesMessage';
